refactor, move sections to smaller components
