+++ packages/atlas/src/views/global/YppLandingView/YppFooter.styles.ts
@@ -1,0 +1,53 @@
+import styled from '@emotion/styled'
+
+import bottomLeftPattern from '@/assets/images/ypp-background-pattern.svg'
+import topLeftBannerPattern from '@/assets/images/ypp-banner-pattern.svg'
+import { Text } from '@/components/Text'
+import { Button } from '@/components/_buttons/Button'
+import { cVar, media, sizes } from '@/styles'
+
+export const CtaBanner = styled.div`
+  padding: ${sizes(16)} ${sizes(8)};
+  background: ${cVar('colorBackgroundPrimary')};
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: ${sizes(16)} 0;
+  ${media.sm} {
+    margin: ${sizes(24)} 0;
+    background-image: url(${bottomLeftPattern}), url(${topLeftBannerPattern});
+    background-position: bottom left, top right;
+    background-repeat: no-repeat, no-repeat;
+  }
+  ${media.md} {
+    padding: ${sizes(24)} ${sizes(12)};
+  }
+`
+
+export const StyledBannerText = styled(Text)`
+  max-width: 400px;
+  ${media.md} {
+    max-width: 500px;
+  }
+  ${media.lg} {
+    max-width: unset;
+  }
+`
+
+export const StyledButton = styled(Button)`
+  margin-top: ${sizes(8)};
+  background-color: ${cVar('colorCoreBaseBlack')};
+`
+
+export const CtaCardRow = styled.div`
+  display: grid;
+  gap: ${sizes(4)};
+  padding-bottom: ${sizes(16)};
+
+  ${media.md} {
+    grid-template-columns: 1fr 1fr 1fr;
+    gap: ${sizes(6)};
+    padding-bottom: ${sizes(24)};
+  }
+`
