+++ packages/atlas/src/views/global/YppLandingView/YppHero.tsx
@@ -16,7 +16,7 @@ import { SvgActionChevronR } from '@/components/_icons'
 import { useMediaMatch } from '@/hooks/useMediaMatch'
 
 import { BackImage, FrontImage, HeroImageWrapper } from './YppHero.styles'
-import { CenteredLimidtedWidthContainer } from './YppLandingView.styles'
+import { BackgroundContainer, CenteredLimidtedWidthContainer } from './YppLandingView.styles'
 
 export const YppHero: FC = () => {
   const mdMatch = useMediaMatch('md')
@@ -28,31 +28,33 @@ export const YppHero: FC = () => {
     translateY: [0, -15],
   })
   return (
-    <CenteredLimidtedWidthContainer>
-      <LayoutGrid>
-        <GridItem as="header" colSpan={{ base: 12, sm: 8, lg: 6 }} colStart={{ sm: 3, lg: 4 }}>
-          <Text as="h1" variant={mdMatch ? 'h800' : 'h600'}>
-            Connect your YouTube channel & get paid
-          </Text>
-          <Text as="p" variant="t300" color="colorText" margin={{ top: 4, bottom: 8 }}>
-            Reupload and backup your YouTube videos to receive a guaranteed payout in the YouTube Partner Program.
-          </Text>
-          <Button size="large" icon={<SvgActionChevronR />} iconPlacement="right">
-            Sign up now
-          </Button>
-          <Text as="p" variant="t100" color="colorText" margin={{ top: 2 }}>
-            It takes 3 minutes and is 100% free.
-          </Text>
-        </GridItem>
-      </LayoutGrid>
-      <HeroImageWrapper>
-        <BackImage srcSet={`${yt576} 576w, ${yt864} 864w, ${yt1152} 1152w, ${yt2304} 2304w`} alt="Hero back" />
-        <FrontImage
-          ref={heroImageRef}
-          srcSet={`${hero576} 576w, ${hero864} 864w, ${hero1152} 1152w, ${hero2304} 2304w`}
-          alt="Hero front"
-        />
-      </HeroImageWrapper>
-    </CenteredLimidtedWidthContainer>
+    <BackgroundContainer noBackground>
+      <CenteredLimidtedWidthContainer>
+        <LayoutGrid>
+          <GridItem as="header" colSpan={{ base: 12, sm: 8, lg: 6 }} colStart={{ sm: 3, lg: 4 }}>
+            <Text as="h1" variant={mdMatch ? 'h800' : 'h600'}>
+              Connect your YouTube channel & get paid
+            </Text>
+            <Text as="p" variant="t300" color="colorText" margin={{ top: 4, bottom: 8 }}>
+              Reupload and backup your YouTube videos to receive a guaranteed payout in the YouTube Partner Program.
+            </Text>
+            <Button size="large" icon={<SvgActionChevronR />} iconPlacement="right">
+              Sign up now
+            </Button>
+            <Text as="p" variant="t100" color="colorText" margin={{ top: 2 }}>
+              It takes 3 minutes and is 100% free.
+            </Text>
+          </GridItem>
+        </LayoutGrid>
+        <HeroImageWrapper>
+          <BackImage srcSet={`${yt576} 576w, ${yt864} 864w, ${yt1152} 1152w, ${yt2304} 2304w`} alt="Hero back" />
+          <FrontImage
+            ref={heroImageRef}
+            srcSet={`${hero576} 576w, ${hero864} 864w, ${hero1152} 1152w, ${hero2304} 2304w`}
+            alt="Hero front"
+          />
+        </HeroImageWrapper>
+      </CenteredLimidtedWidthContainer>
+    </BackgroundContainer>
   )
 }
